working on edits

+++ app/controllers/bios_controller.rb
@@ -34,6 +34,7 @@ class BiosController < ApplicationController
 
   # GET /bios/1/edit
   def edit
+    @images  = Photo.where(bio_id: @bio.id)
   end
 
   # POST /bios
@@ -61,6 +62,11 @@ class BiosController < ApplicationController
       if @bio.update(bio_params)
         save_images_to_s3 params['bio']['images']
         @bio.save
+        params['images'].each do |img|
+          if img.should_remove
+            Photo.find_by(id: img.id).delete
+          end
+        end
         format.html { redirect_to @bio, notice: 'Bio was successfully updated.' }
         format.json { render :show, status: :ok, location: @bio }
       else
@@ -89,7 +95,7 @@ class BiosController < ApplicationController
       image.destroy
 
     end
-    
+
     @bio.destroy
     respond_to do |format|
       format.html { redirect_to bios_url, notice: 'Bio was successfully destroyed.' }
@@ -105,34 +111,37 @@ class BiosController < ApplicationController
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def bio_params
-      params.require(:bio).permit(:name, :age, :breed, :likes, :fears, :details)
+      params.require(:bio).permit(:name, :age, :breed, :likes, :fears, :details, :images)
     end
 
     def save_images_to_s3(image_list)
-      s3 = Aws::S3::Resource.new(region: ENV['S3_REGION'])
-      bucket = s3.bucket(ENV['S3_BUCKET_NAME'])
-      image_list.each do |uf|
-        image = Photo.new
-        s3image = bucket.object("#{@bio.id}_#{Pathname.new(uf.path).basename.to_s}")
-        s3thumb = bucket.object("thumb_#{@bio.id}_#{Pathname.new(uf.path).basename.to_s}")
-        img = Image.read(uf.path)[0]
-        img.auto_orient!
-        img.resize_to_fit!(500, 500)
-        s3image.put(body: img.to_blob, acl: 'public-read')
-        image.path = s3image.public_url
-        image.thumbnail = s3thumb.public_url
-        target = Image.new(100, 100) do
-          self.background_color = 'white'
+      if image_list
+
+        s3 = Aws::S3::Resource.new(region: ENV['S3_REGION'])
+        bucket = s3.bucket(ENV['S3_BUCKET_NAME'])
+        image_list.each do |uf|
+          image = Photo.new
+          s3image = bucket.object("#{@bio.id}_#{Pathname.new(uf.path).basename.to_s}")
+          s3thumb = bucket.object("thumb_#{@bio.id}_#{Pathname.new(uf.path).basename.to_s}")
+          img = Image.read(uf.path)[0]
+          img.auto_orient!
+          img.resize_to_fit!(500, 500)
+          s3image.put(body: img.to_blob, acl: 'public-read')
+          image.path = s3image.public_url
+          image.thumbnail = s3thumb.public_url
+          target = Image.new(100, 100) do
+            self.background_color = 'white'
+          end
+          target.format = img.format
+          img.resize_to_fit!(100, 100)
+          s3thumb.put(body: target.composite(img, CenterGravity, CopyCompositeOp).to_blob, acl: 'public-read')
+          if @bio.main_image == nil
+            @bio.main_image = s3thumb.public_url
+          end
+          @bio.photos << image
+          image.save
+          uf.tempfile.unlink
         end
-        target.format = img.format
-        img.resize_to_fit!(100, 100)
-        s3thumb.put(body: target.composite(img, CenterGravity, CopyCompositeOp).to_blob, acl: 'public-read')
-        if @bio.main_image == nil
-          @bio.main_image = s3thumb.public_url
-        end
-        @bio.photos << image
-        image.save
-        uf.tempfile.unlink
       end
     end
 end
